requirements.txt
# Core ML
pandas>=2.1.0
numpy>=1.24.3
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Model Explainability
shap>=0.43.0

# Streamlit
streamlit>=1.28.0

# GenAI & LLM
langchain>=0.1.0
langchain-google-genai>=0.0.6

# Vector Database
chromadb>=0.4.18

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.2
